+++ src/main/java/com/github/cc3002/finalreality/controller/IEventHandler.java
@@ -1,0 +1,6 @@
+package com.github.cc3002.finalreality.controller;
+
+import java.beans.PropertyChangeListener;
+
+public interface IEventHandler extends PropertyChangeListener {
+}