+++ src/main/java/com/github/cc3002/finalreality/model/character/player/Knight.java
@@ -1,7 +1,6 @@
 package com.github.cc3002.finalreality.model.character.player;
 
 import com.github.cc3002.finalreality.model.character.ICharacter;
-import com.github.cc3002.finalreality.model.weapon.AbstractWeapon;
 import com.github.cc3002.finalreality.model.weapon.IWeapon;
 import org.jetbrains.annotations.NotNull;
 
@@ -10,12 +9,12 @@ import java.util.concurrent.BlockingQueue;
 
 public class Knight extends PlayerCharacter{
     /**
-     * Creates a new common character.
+     * Creates a new Knight character.
      *
      * @param name       the common character's name
      * @param points     the common character's points
      * @param defense    the common character's defense
-     * @param turnsQueue
+     * @param turnsQueue the queue with the characters waiting for their turn
      */
     public Knight(String name, int points, int defense,  @NotNull BlockingQueue<ICharacter> turnsQueue) {
         super(name, points, defense, turnsQueue);
@@ -23,7 +22,9 @@ public class Knight extends PlayerCharacter{
 
     @Override
     public void equip(IWeapon weapon){
+        if(this.isAlive()){
         weapon.equipToKnight(this);
+        }
     }
 
     @Override