+++ src/test/java/com/github/cc3002/finalreality/model/character/AbstractCharacterTest.java
@@ -1,11 +1,8 @@
 package com.github.cc3002.finalreality.model.character;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;