+++ src/test/java/com/github/cc3002/finalreality/model/character/AbstractPlayerCharacterTest.java
@@ -1,13 +1,7 @@
 package com.github.cc3002.finalreality.model.character;
 
 import com.github.cc3002.finalreality.model.character.player.IPlayerCharacter;
-
-import com.github.cc3002.finalreality.model.character.player.Knight;
-import com.github.cc3002.finalreality.model.character.player.Thief;
 import com.github.cc3002.finalreality.model.weapon.*;
-import org.junit.jupiter.api.Test;
-
-import java.util.concurrent.LinkedBlockingQueue;
 
 import static org.junit.jupiter.api.Assertions.*;